fix(docs): no extraneous exports from route page
